optimize schema
